% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outage_table.R
\name{outage_table}
\alias{outage_table}
\title{Convolve generators to create outage tables}
\usage{
outage_table(generators, levels = NULL, threshold = 1e-06, round = 1)
}
\arguments{
\item{generators}{Data frame with the list of generators (see details for requirements)}

\item{levels}{Optional data frame that contains levels of aggregation, e.g., BAA, transmission
area or interconnection (see details for requirements)}

\item{threshold}{Table entries with LOLP below this value are ignored}

\item{round}{Used to round capacities (default is 1 MW)}
}
\value{
The result is a \link[data.table]{data.table} object that is used to calculate
LOLP metrics. For each area, the output presents the following columns: \code{Capacity},
\code{Prob}, \code{LOLP} and \code{BaseEUE}. The column \code{Capacity} is
duplicated as (\code{Capacity2}) and is used internally by the LOLP calculation functions.

Given a load level \code{L}, the probability parameters are calculated as follows:
\itemize{
  \item{Choose the first row with \code{L <= Capacity}}
  \item{The loss-of-load probability is the value in the \code{LOLP} for that row}
  \item{Expected unserved energy can be quickly calculated as
        \code{EUE = BaseEUE + (L - Capacity) * LOLP}. This is
        equivalent to calculating the sum of \code{(L - Capacity) * Prob} for all
        rows with \code{L < Capacity}}
}
}
\description{
Allows to create outage tables for different areas and levels of aggregation for a list of capacities
and equivalent forced outage rates (EFOR) for generators. The assumption is that the
generator outages are independent from each other. The table can be used
to quickly compute loss-of-load probability (LOLP) and expected unserved
energy (EUE).
}
\details{
Requirements for \code{generators}:
\itemize{
  \item{The data frame must contain two columns: \code{Capacity} and \code{EFOR} (Other columns will
        be ignored)}
  \item{Optionally if a column called \code{Area} exists, it will be used to create separate tables
        within each area}
}

Requirements for \code{levels}:
\itemize{
  \item{This parameter is optional}
  \item{If \code{levels} is provided, \code{generators} must contain a column called \code{Area}}
  \item{The first column must contain the most granular level of data and compatible with columns
        \code{Area} in \code{generators}}
  \item{The column names in \code{levels} will become the names of the different levels of aggregation}
}
}
\examples{
# Outage table with one 10-MW generator with 2\% EFOR
outage_table(data.frame(Capacity = 10, EFOR = 0.02))

# Outage table with two generators (10 MW with 2\% EFOR and 20 MW with 1\% EFOR)
outage_table(data.frame(Capacity = c(10, 20), EFOR = c(0.02, 0.01)))

# List of generators and areas
gens <- data.frame(Area = c(rep("A", 10), rep("B", 5)),
                   Capacity = rep(60, 15),
                   EFOR = rep(0.08, 15))
levs <- data.frame(BAA = c("A", "B"), Region = c("All", "All"))

# Create a single outage table (without specifying area)
outage_table(gens[, c("Capacity", "EFOR")])

# Create an outage table for each 'Area'
outage_table(gens)

# Create table for each 'Area' and different and levels of aggregation in 'levs'
outage_table(gens, levs)
}
\references{
R. Billinton, and R. N. Allan, Reliability evaluation of power systems,
    New York: Plenum Press 1996.
}
\seealso{
\code{\link{format_timedata}} is the function that creates compatible time data objects
}

