% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{repra-package}
\alias{repra}
\alias{repra-package}
\title{The Renewable Energy Probabilistic Resource Assessment tool}
\description{
This package is the R implementation of the Renewable Energy Probabilistic Resource Assessment (REPRA)
tool, developed at the National Renewable Energy Laboratory (NREL).
}
\details{
More information is available in the help of the package functions. Additionally, two vignettes
are included with this package and can be accessed with \code{browseVignettes("repra")} or
\code{browseVignettes("sliding_window")}.
}
\references{
Eduardo Ibanez, Michael Milligan (2012). Impact of Transmission on Resource
  Adequacy in Systems with Wind and Solar Power. IEEE Power& Energy Society General
  Meeting, 22-26 July 2012, San Diego, CA.
  \url{http://www.nrel.gov/docs/fy12osti/53482.pdf}.
}

